Frontend layout refactor
