+++ website/src/assets/main.css
@@ -16,6 +16,38 @@ a,
   vertical-align: center;
 }
 
+main {
+  width: 98%;
+  margin-top: 10rem;
+  margin-left: 1%;
+  justify-content: left;
+  min-height: 100%;
+  padding: 1% 1% 0 1%;
+  position: relative;
+  float: left;
+  margin-bottom: 4rem;
+}
+
+section {
+  width: 17%;
+  background-color: rgba(240, 248, 255, 0.2);
+  top: 0;
+  /* margin-top: 7.5rem; */
+  margin-right: 1%;
+  justify-content: center;
+  min-height: 100%;
+  position: relative;
+  padding: 0 1% 0 1%;
+  float: left;
+  font-size: 1vmax;
+}
+
+.section_title {
+  font-size: 2rem;
+  font-weight: 500;
+  margin-bottom: -1%;
+}
+
 @media (min-width: 1024px) {
   body {
     display: flex;
@@ -23,3 +55,4 @@ a,
     margin: 0;
   }
 }
+
