--- website/src/components/Business.vue
@@ -1,12 +1,0 @@
-<script setup>
-import Currencies from "@/components/Currencies.vue";
-import TheMainContent from "@/components/TheMainContent.vue";
-import ArticleSection from "./ArticleSection.vue";
-
-</script>
-
-<template>
-
-    <Currencies/>
-    <ArticleSection/>
-</template>
