+++ website/src/components/api-currencies/Currencies.vue
@@ -9,7 +9,7 @@
 <script>
 import { ref } from "vue";
 import SlidingArray from "./SlidingArray.vue";
-import { useCurrencies } from './CurrenciesService'
+import { useCurrencies } from '@/scripts/CurrenciesService.ts'
 
 export default {
   components: {
