+++ website/src/components/api-currencies/SlidingArray.vue
