+++ website/src/components/articles/Article.vue
@@ -2,7 +2,7 @@
 import { QuillEditor } from '@vueup/vue-quill'
 import '@vueup/vue-quill/dist/vue-quill.snow.css';
 import '@vueup/vue-quill/dist/vue-quill.core.css';
-import { ref, onMounted, defineProps} from 'vue'
+import { ref, onMounted} from 'vue'
 import { useRoute } from 'vue-router'
 
 
@@ -16,8 +16,8 @@ const props = defineProps({
 onMounted(() => {
     console.log(route.query["id"]);
     let index = route.query["id"]
-    console.log(JSON.parse(localStorage.getItem("allArticles") || [])[index])
-    content.value = JSON.parse(localStorage.getItem("allArticles") || [])[index].content
+    console.log(JSON.parse(sessionStorage.getItem("allArticles") || [])[index])
+    content.value = JSON.parse(sessionStorage.getItem("allArticles") || [])[index].content
 })
 
 </script>
