+++ website/src/components/user-layout/UserInfo.vue
@@ -1,0 +1,9 @@
+<script setup>
+import jsCookie from 'js-cookie';
+</script>
+
+<template>
+
+    <p v-html="jsCookie.get('role')"></p>
+
+</template>
