+++ website/src/data/login.json
@@ -1,6 +1,6 @@
 [
   {
     "link":"/login",
-    "text":"Zaloguj/Zarejestrj"
+    "text":"Zaloguj/Zarejestruj"
   }
 ]
