+++ website/src/scripts/CurrenciesModel.d.ts
