+++ website/src/scripts/CurrenciesService.ts
