+++ website/src/scripts/WeatherModel.ts
